Mejoras y cambio de interfaz en mover.sh

+++ mover.sh
@@ -1,28 +1,38 @@
+function getCodigoDuplicado {
+	CODIGO_DUP=0
+	for i in $(ls -F "$1" | grep "$2" | sed 's/.*\-//g' | sed 's/\..*//');
+	do
+		if [[ ! $i == *"_"* ]]; then #Es el primer archivo, no tiene codigo agregado aun
+			if [ $i -gt $CODIGO_DUP ]; then
+				CODIGO_DUP=$i;
+			fi
+		fi
+	done;
+	CODIGO_DUP=$((CODIGO_DUP + 1));
+	return $CODIGO_DUP
+}
+
 #Script auxiliar para mover archivos, teniendo en cuenta el caso en el que ya exista un archivo con el mismo nombre en la carpeta destino
 #Recibe cinco parámetros
 
 origen=$1 #Incluye el nombre del archivo
 destino=$2 #No incluye el nombre del archivo
-nombre_archivo=$3
-alternativa=$4 #Ruta alternativa a donde copiar si ya existe el archivo en el destino
+alternativa=$3 #Ruta alternativa a donde copiar si ya existe el archivo en el destino
 
+nombre_archivo=${origen##*/}
 sin_extension=${nombre_archivo%.*}
 extension=${nombre_archivo##*.}
-
-#Obtengo el codigo a agregar a los archivos repetidos para no perderlos
-codigo=1
-if [ ! -f "$EJECUTABLES""/hash.txt" ]; then
-	echo "$codigo" > "$EJECUTABLES""/hash.txt"
-else
-	codigo=$(cat "$EJECUTABLES""/hash.txt")
-fi
-
 
 if [ ! -f "$destino"/"$nombre_archivo" ]; then #Si no existe el archivo, lo muevo a la carpeta destino
 	mv "$origen" "$destino"/"$nombre_archivo"
 else
 	mkdir -p "$alternativa" #Si no existe, creo el directorio de duplicados
-	mv "$origen" "$alternativa"/"$sin_extension""-""$codigo"".""$extension"
-	let codigo+=1 #Actualizo el valor para que no se repitan dos archivos
-	echo "$codigo" > "$EJECUTABLES""/hash.txt"
+	#Obtengo el codigo a agregar a los archivos repetidos para no perderlos
+	if [ -f "$alternativa"/"$nombre_archivo" ]; then
+		getCodigoDuplicado "$alternativa" "$sin_extension"
+		codigo=$?
+		mv "$origen" "$alternativa"/"$sin_extension""-""$codigo"".""$extension"
+	else
+		mv "$origen" "$alternativa"/"$nombre_archivo"
+	fi
 fi
